ES6 support with depcheck-es6
If your code has JSX, or uses stuff that acorn or node-soure-walker cannot
parse, such as `static`, then it's still going to say that you haven't included
something.

Not sure what the right course of action is, as this is really sort of an
interesting problem for the parsing libraries that we use.

Anyway, if you're using plain ol' ES6 without anything fancy, this should
resolve your imports for you just fine.